resume_matcher_mvp/requirements.txt
gradio==4.44.0
sentence-transformers==2.2.2
chromadb==0.4.15
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
